Split up Input into Input and InputHandler
